pyproject.toml
[project]
name = "agents-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
